+++ BALSAMIC/commands/config/base.py
@@ -2,7 +2,7 @@
 import click
 
 from BALSAMIC.commands.config.case import case_config as case_command
-
+from BALSAMIC.commands.config.pon import pon_config as pon_command
 
 @click.group()
 @click.pass_context
@@ -12,3 +12,4 @@ def config(context):
 
 
 config.add_command(case_command)
+config.add_command(pon_command)