+++ BALSAMIC/utils/models.py
@@ -45,10 +45,10 @@ class VarCallerFilter(BaseModel):
         MQ: VCFAttributes (optional); minimum mapping quality
         DP: VCFAttributes (optional); minimum read depth
         pop_freq: VCFAttributes (optional); maximum gnomad_af
-	    strand_reads: VCFAttributes (optional); minimum strand specific read counts
+        strand_reads: VCFAttributes (optional); minimum strand specific read counts
         qss: VCFAttributes (optional); minimum sum of base quality scores
-	    sor: VCFAttributes (optional); minimum symmetrical log-odds ratio
-	    varcaller_name: str (required); variant caller name
+        sor: VCFAttributes (optional); minimum symmetrical log-odds ratio
+        varcaller_name: str (required); variant caller name
         filter_type: str (required); filter name for variant caller
         analysis_type: str (required); analysis type e.g. tumor_normal or tumor_only
         description: str (required); comment section for description
@@ -168,30 +168,28 @@ class AnalysisModel(BaseModel):
     """Pydantic model containing workflow variables
 
     Attributes:
-    
+              
         case_id : Field(required); string case identifier
-        analysis_type : Field(required); string literal [single, paired]
+        analysis_type : Field(required); string literal [single, paired, pon]
             single : if only tumor samples are provided
             paired : if both tumor and normal samples are provided
+            pon : panel of normal analysis
         sequencing_type : Field(required); string literal [targeted, wgs]
             targeted : if capture kit was used to enrich specific genomic regions
             wgs : if whole genome sequencing was performed
         analysis_dir : Field(required); existing path where to save files
-        umiworkflow : Field(bool); whether UMI workflow to run parallely 
-
         fastq_path : Field(optional); Path where fastq files will be stored
         script : Field(optional); Path where snakemake scripts will be stored
         log : Field(optional); Path where logs will be saved
         result : Field(optional); Path where BALSAMIC output will be stored
         benchmark : Field(optional); Path where benchmark report will be stored
         dag : Field(optional); Path where DAG graph of workflow will be stored
-
         BALSAMIC_version  : Field(optional); Current version of BALSAMIC
         config_creation_date  : Field(optional); Timestamp when config was created
 
     Raises:
         ValueError:
-            When analysis_type is set to any value other than [single, paired, qc]
+            When analysis_type is set to any value other than [single, paired, qc, pon]
             When sequencing_type is set to any value other than [wgs, targeted]
     """
 
@@ -207,7 +205,6 @@ class AnalysisModel(BaseModel):
     dag: Optional[FilePath]
     BALSAMIC_version: str = balsamic_version
     config_creation_date: Optional[str]
-    umiworkflow: bool
 
     class Config:
         validate_all = True
@@ -324,21 +321,53 @@ class PanelModel(BaseModel):
         return Path(value).resolve().as_posix()
 
 
+class PonBalsamicConfigModel(BaseModel):
+    """Summarizes config models in preparation for export
+
+    Attributes:
+        QC : Field(QCmodel); variables relevant for fastq preprocessing and QC
+        reference : Field(Dict); dictionary containing paths to reference genome files
+        panel : Field(PanelModel(optional)); variables relevant to PANEL BED if capture kit is used
+        singularity : Field(Path); path to singularity container of BALSAMIC
+        rule_directory : Field(Path(RULE_DIRECTORY)); path where snakemake rules can be found
+        bioinfo_tools : Field(dict); dictionary of bioinformatics software and which conda/container they are in
+        bioinfo_tools_version : Field(dict); dictionary of bioinformatics software and their versions used for the analysis
+    """
+
+    QC: QCModel
+    analysis: AnalysisModel
+    reference: Dict[str, Path]
+    singularity: DirectoryPath
+    bioinfo_tools: dict
+    bioinfo_tools_version: dict
+    panel: Optional[PanelModel]
+
+    @validator("reference")
+    def abspath_as_str(cls, value):
+        for k, v in value.items():
+            value[k] = Path(v).resolve().as_posix()
+        return value
+
+    @validator("singularity")
+    def transform_path_to_dict(cls, value):
+        return {"image": Path(value).resolve().as_posix()}
+
+
 class BalsamicConfigModel(BaseModel):
     """Summarizes config models in preparation for export 
     
     Attributes:
         QC : Field(QCmodel); variables relevant for fastq preprocessing and QC
-        vcf : Field(VCFmodel); variables relevand for variant calling pipeline
+        vcf : Field(VCFmodel); variables relevant for variant calling pipeline
         samples : Field(Dict); dictionary containing samples submitted for analysis
-        reference : Field(Dict); dictionary containign paths to reference genome files
+        reference : Field(Dict); dictionary containing paths to reference genome files
         panel : Field(PanelModel(optional)); variables relevant to PANEL BED if capture kit is used
         bioinfo_tools : Field(dict); dictionary of bioinformatics software and which conda/container they are in
         bioinfo_tools_version : Field(dict); dictionary of bioinformatics software and their versions used for the analysis
         singularity : Field(Path); path to singularity container of BALSAMIC
         background_variants: Field(Path(optional)); path to BACKGROUND VARIANTS for UMI
         rule_directory : Field(Path(RULE_DIRECTORY)); path where snakemake rules can be found
-
+        umiworkflow : Field(bool); whether UMI workflow to run in parallel
     """
 
     QC: QCModel
@@ -351,6 +380,7 @@ class BalsamicConfigModel(BaseModel):
     bioinfo_tools: dict
     bioinfo_tools_version: dict
     panel: Optional[PanelModel]
+    umiworkflow: bool
 
     @validator("reference")
     def abspath_as_str(cls, value):
@@ -435,7 +465,7 @@ class ReferenceMeta(BaseModel):
     This class defines a meta for various reference files. Only reference_genome is mandatory.
 
     Attributes:
-      basedir: str for basedirectory which will be appended to all ReferenceUrlsModel fields
+      basedir: str for base directory which will be appended to all ReferenceUrlsModel fields
       reference_genome: ReferenceUrlsModel. Required field for reference genome fasta file
       dbsnp: ReferenceUrlsModel. Optional field for dbSNP vcf file
       hc_vcf_1kg: ReferenceUrlsModel. Optional field for high confidence 1000Genome vcf
@@ -533,7 +563,7 @@ class UMIParamsTNscope(BaseModel):
         filter_tumor_af: float (required); minimum allelic frequency to detect
         min_tumorLOD: int (required); minimum tumor log odds in the final call of variants
         init_tumorLOD: float (required); minimum tumor log odds in the initial pass calling variants
-	    error_rate: int (required); allow error-rate to consider in calling
+        error_rate: int (required); allow error-rate to consider in calling
         prunefactor: int (required); pruning factor in the kmer graph
     """