+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Added:
 
 * Changelog reminder workflow to Github
 * Snakemake workflow for created PON reference
+* Balsamic cli config command(pon) for creating json for PON analysis
 * Added tumor lod option for passing tnscope-umi final variants